requirements.txt
mcp>=1.0.0
mysql-connector-python>=9.1.0
pypinyin>=0.48.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
starlette>=0.27.0
jieba~=0.42.1
scikit-learn~=1.6.1
neo4j~=5.28.1
nltk~=3.9.1
sentence-transformers~=4.1.0